+++ test/typings.test.ts
@@ -661,6 +661,11 @@ describe('typings.test.js', function () {
                 const myDb: RxDatabase = {} as any;
                 const typedLocalDoc = await myDb.getLocal<{foo: string;}>('foobar');
                 const typedLocalDocInsert = await myDb.insertLocal<{foo: string;}>('foobar', { bar: 'foo' });
+
+                if (!typedLocalDoc) {
+                    throw new Error('local doc missing');
+                }
+
                 const x: string = typedLocalDoc.foo;
                 const x2: string = typedLocalDocInsert.foo;
             });
@@ -676,6 +681,11 @@ describe('typings.test.js', function () {
                 const myDb: RxDatabase = {} as any;
                 const typedLocalDoc = await myDb.getLocal<{foo: string;}>('foobar');
                 const typedLocalDocUpsert = await myDb.upsertLocal<{foo: string;}>('foobar', { foo: 'bar' });
+
+                if (!typedLocalDoc) {
+                    throw new Error('local doc missing');
+                }
+
                 const x: string = typedLocalDoc.foo;
                 const x2: string = typedLocalDocUpsert.foo;
             });